requirements.txt
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
attrs==22.1.0
Authlib==1.2.1
beautifulsoup4==4.11.2
blessed==1.19.1
boto3==1.17.81
botocore==1.20.112
cbor2==5.4.3
certifi==2022.12.7
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.2
Django==4.2.6
django-bootstrap5==22.2
django-cors-headers==4.4.0
django-picklefield==3.1
django-q==1.3.9
django-tables2==2.5.1
ecdsa==0.18.0
idna==3.4
intel-sgx-ra==1.0.1
isodate==0.6.1
jmespath==0.10.0
oscrypto==1.3.0
pycose==1.0.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.31.0
s3transfer==0.4.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.8.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
